move usecase to src/hackingBuddyGPT
