+++ src/hackingBuddyGPT/capabilities/__init__.py
@@ -1,4 +1,4 @@
-from .capability import Capability
+from ..capability import Capability
 from .psexec_run_command import PSExecRunCommand
 from .psexec_test_credential import PSExecTestCredential
 from .ssh_run_command import SSHRunCommand
