+++ src/hackingBuddyGPT/capabilities/psexec_test_credential.py
@@ -4,7 +4,7 @@ from typing import Tuple
 
 from hackingBuddyGPT.utils import PSExecConnection
 
-from .capability import Capability
+from ..capability import Capability
 
 
 @dataclass
