+++ src/hackingBuddyGPT/capability.py
