+++ src/hackingBuddyGPT/usecases/agents.py
@@ -5,7 +5,7 @@ from mako.template import Template
 from typing import Dict
 
 from hackingBuddyGPT.utils.logging import log_conversation, Logger, log_param
-from hackingBuddyGPT.capabilities.capability import (
+from hackingBuddyGPT.capability import (
     Capability,
     capabilities_to_simple_text_handler,
 )
